+++ src/main/java/com/codeup/sidework/controllers/UserController.java
@@ -1,6 +1,5 @@
 package com.codeup.sidework.controllers;
 
-import com.codeup.sidework.models.Business;
 import com.codeup.sidework.models.User;
 import com.codeup.sidework.repositories.Users;
 import com.codeup.sidework.repositories.UsersRepository;
@@ -10,7 +9,6 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Arrays;
-
 
 @Controller
 public class UserController {
@@ -24,20 +22,19 @@ public class UserController {
         this.passwordEncoder = passwordEncoder;
     }
 
-    @GetMapping("/users/register-mgmt")
-    public String showManagementRegisterForm(Model model) {
-        model.addAttribute("user", new User());
-        model.addAttribute("business", new Business());
-        return "users/register-mgmt";
-    }
-
-    @PostMapping("/register")
-    public String saveManagement(@ModelAttribute User user) {
-        String hash = passwordEncoder.encode(user.getPassword());
-        user.setPassword(hash);
-        users.save(user);
-        return "redirect:/users/login-mgmt";
-    }
+//    @GetMapping("/users/register-mgmt")
+//    public String showManagementRegisterForm(Model model) {
+//        model.addAttribute("user", new User());
+//        return "users/register-mgmt";
+//    }
+//
+//    @PostMapping("/register")
+//    public String saveManagement(@ModelAttribute User user) {
+//        String hash = passwordEncoder.encode(user.getPassword());
+//        user.setPassword(hash);
+//        users.save(user);
+//        return "redirect:/users/login-mgmt";
+//    }
 
     @GetMapping("/users/register-worker")
     public String showWorkerRegisterForm(Model model) {
@@ -51,8 +48,10 @@ public class UserController {
                              @RequestParam(value = "employed") boolean isEmployed,
                              @RequestParam(value = "available") boolean isAvailable) {
         String hash = passwordEncoder.encode(user.getPassword());
+
         user.setPassword(hash);
         user.setSkills(Arrays.toString(checkboxValue));
+
         if (isEmployed) {
             user.setCurrentEmployment(true);
         } else {
@@ -68,11 +67,6 @@ public class UserController {
         users.save(user);
 
         return "redirect:/users/login-worker";
-    }
-
-    @GetMapping("/users/workspace-mgmt")
-    public String showManagementWorkspace() {
-        return "users/workspace-mgmt";
     }
 
     @GetMapping("/users/workspace-worker")
