+++ src/main/resources/templates/users/register-mgmt.html
@@ -33,10 +33,10 @@
     <div class="container">
         <div class="pic"></div>
         <h1>register</h1>
-        <form th:action="@{/register}" method="post" th:object="${user}">
+        <form th:action="@{/register}" method="post" th:object="${business}">
             <div class="form-group">
-                <label for="username">username</label>
-                <input id="username" th:field="*{username}" class="form-control" type="text"/>
+                <label for="username">business manager username</label>
+                <input id="username" th:field="*{business_manager}" class="form-control" type="text"/>
             </div>
             <div class="form-group">
                 <label for="email">email</label>
@@ -50,8 +50,6 @@
                 <label for="confirm_password">confirm password</label>
                 <input id="confirm_password" name="confirm_password" class="form-control" type="password"/>
             </div>
-        </form>
-        <form th:action="@{/registerbusiness}" method="post" th:object="${business}">
             <div class="form-group">
                 <label for="business_name">business name</label>
                 <input id="business_name" th:field="*{business_name}" class="form-control" type="text"/>
